pyproject.toml
[build-system]
requires = ["setuptools >= 35.0.2", "wheel >= 0.29.0", "poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.isort]
combine_as_imports = true
default_section = "THIRDPARTY"
force_grid_wrap = 0
force_sort_within_sections = true
forced_separate = "tests"
include_trailing_comma = true
indent = "    "
known_first_party = "pybhyve,tests"
line_length = 88
multi_line_output = 3
not_skip = "__init__.py"
sections = "FUTURE,STDLIB,INBETWEENS,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
use_parentheses = true

[tool.poetry]
name = "pybhyve"
version = "1.0.2"
description = "A clean, async-friendly library for the Orbit BHyve API"
readme = "README.md"
authors = ["Seb Ruiz <seb@sebruiz.net>"]
license = "MIT"
repository = "https://github.com/sebr/pybhyve"
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]

[tool.poetry.dependencies]
aiohttp = "^3.6.2"
python = "^3.6.1"
python-engineio = "=3.10.0"
python-socketio = "=4.4.0"
websockets = "^8.1"

[tool.poetry.dev-dependencies]
aresponses = "^1.1.1"
bandit = "^1.6.2"
black = "^19.10b0"
flake8 = "^3.7.9"
mypy = "^0.740"
pre-commit = "^1.20.0"
pydocstyle = "^4.0.1"
pylint = "^2.4.3"
pytest-aiohttp = "^0.3.0"
pytest-cov = "^2.8.1"
